Clean up (#16)
* refactor: move directories, get license for header

* refactor: use licence as header in files

* refactor: rustfmt world

* refactor: clean-up

* fix: compilation and tests after refactor

* fix: integration-tests and compilation

* feat: add workspace test to workflow
