+++ .github/workflows/build.yml
@@ -26,4 +26,4 @@ jobs:
       - name: build
         run: cargo build --release
       - name: test
-        run: cargo test --release
+        run: cargo test --workspace --release
