+++ Cargo.toml
@@ -1,2 +1,2 @@
 [workspace]
-members = ["gateway", "libs/ecdsa", "integration-test", "sample-runtime", "sample-receiver"]
+members = ["gateway", "libs/ecdsa", "integration-test", "integration-test/sample-runtime", "integration-test/sample-receiver"]
