+++ gateway/src/proof.rs
@@ -1,134 +1,144 @@
+// Copyright 2023 Centrifuge Foundation (centrifuge.io).
+//
+// This file is part of the axelar-cgp-substrate project.
+// Axelar-cgp-substrate is free software: you can redistribute it and/or modify
+// it under the terms of the GNU General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version (see http://www.gnu.org/licenses).
+// Centrifuge is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU General Public License for more details.
+
 //! # Axelar CGP
 //!
 //! This pallet implements the Axelar Cross-Chain Gateway Protocol
 
-// Ensure we're `no_std` when compiling for WebAssembly.
-#![cfg_attr(not(feature = "std"), no_std)]
 use ethabi::{Address, ParamType, Token};
 use sp_core::{keccak_256, H256};
 
 #[derive(PartialEq, Debug)]
 pub struct Proof {
-    pub operators: Vec<Address>,
-    pub weights: Vec<u128>,
-    pub threshold: u128,
-    pub signatures: Vec<Vec<u8>>,
+	pub operators: Vec<Address>,
+	pub weights: Vec<u128>,
+	pub threshold: u128,
+	pub signatures: Vec<Vec<u8>>,
 }
 
 impl TryFrom<Vec<Token>> for Proof {
-    type Error = ethabi::Error;
+	type Error = ethabi::Error;
 
-    fn try_from(tokens: Vec<Token>) -> Result<Self, Self::Error> {
-        if let [Token::Array(operators_token), Token::Array(weights_token), Token::Uint(t), Token::Array(signatures_token)] =
-            tokens.as_slice()
-        {
-            let operators = operators_token
-                .into_iter()
-                .flat_map(|x| match x {
-                    Token::Address(x) => Ok(x.clone()),
-                    _ => Err(ethabi::Error::InvalidData),
-                })
-                .collect();
-            let weights = weights_token
-                .into_iter()
-                .flat_map(|x| match x {
-                    Token::Uint(w) => Ok(w.as_u128()),
-                    _ => Err(ethabi::Error::InvalidData),
-                })
-                .collect();
-            let threshold = t.as_u128();
-            let signatures = signatures_token
-                .into_iter()
-                .flat_map(|x| match x {
-                    Token::Bytes(x) => Ok(x.clone()),
-                    _ => Err(ethabi::Error::InvalidData),
-                })
-                .collect();
+	fn try_from(tokens: Vec<Token>) -> Result<Self, Self::Error> {
+		if let [Token::Array(operators_token), Token::Array(weights_token), Token::Uint(t), Token::Array(signatures_token)] =
+			tokens.as_slice()
+		{
+			let operators = operators_token
+				.into_iter()
+				.flat_map(|x| match x {
+					Token::Address(x) => Ok(x.clone()),
+					_ => Err(ethabi::Error::InvalidData),
+				})
+				.collect();
+			let weights = weights_token
+				.into_iter()
+				.flat_map(|x| match x {
+					Token::Uint(w) => Ok(w.as_u128()),
+					_ => Err(ethabi::Error::InvalidData),
+				})
+				.collect();
+			let threshold = t.as_u128();
+			let signatures = signatures_token
+				.into_iter()
+				.flat_map(|x| match x {
+					Token::Bytes(x) => Ok(x.clone()),
+					_ => Err(ethabi::Error::InvalidData),
+				})
+				.collect();
 
-            return Ok(Proof {
-                operators,
-                weights,
-                threshold,
-                signatures,
-            });
-        }
+			return Ok(Proof {
+				operators,
+				weights,
+				threshold,
+				signatures,
+			});
+		}
 
-        Err(ethabi::Error::InvalidData)
-    }
+		Err(ethabi::Error::InvalidData)
+	}
 }
 
 /// Obtain the operators set hash by computing the H256 hash of the encoded operators, weights, and threshold.
 pub fn operators_hash(operators: Vec<Address>, weights: Vec<u128>, threshold: u128) -> H256 {
-    let operators_token = operators
-        .into_iter()
-        .map(|x| Token::Address(x.into()))
-        .collect();
-    let weights_token = weights.into_iter().map(|x| Token::Uint(x.into())).collect();
+	let operators_token = operators
+		.into_iter()
+		.map(|x| Token::Address(x.into()))
+		.collect();
+	let weights_token = weights.into_iter().map(|x| Token::Uint(x.into())).collect();
 
-    H256::from(keccak_256(&ethabi::encode(&[
-        Token::Array(operators_token),
-        Token::Array(weights_token),
-        Token::Uint(threshold.into()),
-    ])))
+	H256::from(keccak_256(&ethabi::encode(&[
+		Token::Array(operators_token),
+		Token::Array(weights_token),
+		Token::Uint(threshold.into()),
+	])))
 }
 
 /// Decode a payload expected to contain a `Proof`.
 #[allow(dead_code)]
 pub fn decode(payload: &[u8]) -> Result<Proof, ethabi::Error> {
-    ethabi::decode(
-        &[
-            // operator's addresses
-            ParamType::Array(Box::new(ParamType::Address)),
-            // weights
-            ParamType::Array(Box::new(ParamType::Uint(usize::MAX))),
-            // threshold
-            ParamType::Uint(usize::MAX),
-            // signatures
-            ParamType::Array(Box::new(ParamType::Bytes)),
-        ],
-        payload,
-    )
-    .map(|x| Proof::try_from(x))?
+	ethabi::decode(
+		&[
+			// operator's addresses
+			ParamType::Array(Box::new(ParamType::Address)),
+			// weights
+			ParamType::Array(Box::new(ParamType::Uint(usize::MAX))),
+			// threshold
+			ParamType::Uint(usize::MAX),
+			// signatures
+			ParamType::Array(Box::new(ParamType::Bytes)),
+		],
+		payload,
+	)
+	.map(|x| Proof::try_from(x))?
 }
 
 #[derive(PartialEq, Debug)]
 pub enum SignatureError {
-    /// The signature is invalid
-    InvalidSignature,
-    /// Couldn't find the signer of a signature in the list of operators
-    MalformedSigners,
-    /// Not enough signatures found to meet the threshold
-    LowSignaturesWeight,
+	/// The signature is invalid
+	InvalidSignature,
+	/// Couldn't find the signer of a signature in the list of operators
+	MalformedSigners,
+	/// Not enough signatures found to meet the threshold
+	LowSignaturesWeight,
 }
 
 /// Verifies that the proof holds enough signatures to meet the required threshold.
 /// Fails if not enough operators signed the `msg_hash` to meet the threshold.
 pub fn validate_signatures(msg_hash: H256, proof: Proof) -> Result<(), SignatureError> {
-    let Proof {
-        operators,
-        weights,
-        threshold,
-        signatures,
-    } = proof;
-    let mut weight = 0;
+	let Proof {
+		operators,
+		weights,
+		threshold,
+		signatures,
+	} = proof;
+	let mut weight = 0;
 
-    for signature in signatures.into_iter() {
-        let signer =
-            ecdsa::recover(msg_hash, signature).map_err(|_| SignatureError::InvalidSignature)?;
+	for signature in signatures.into_iter() {
+		let signer =
+			ecdsa::recover(msg_hash, signature).map_err(|_| SignatureError::InvalidSignature)?;
 
-        let index = operators
-            .iter()
-            .position(|x| x.0 == signer.0)
-            .ok_or(SignatureError::MalformedSigners)?;
+		let index = operators
+			.iter()
+			.position(|x| x.0 == signer.0)
+			.ok_or(SignatureError::MalformedSigners)?;
 
-        weight += weights[index];
+		weight += weights[index];
 
-        if weight >= threshold {
-            return Ok(());
-        }
-    }
+		if weight >= threshold {
+			return Ok(());
+		}
+	}
 
-    Err(SignatureError::LowSignaturesWeight)
+	Err(SignatureError::LowSignaturesWeight)
 }
 
 // ----------------------------------------------------------------------------
@@ -136,66 +146,67 @@ pub fn validate_signatures(msg_hash: H256, proof: Proof) -> Result<(), Signature
 // ----------------------------------------------------------------------------
 #[cfg(test)]
 pub mod proof_tests {
-    use super::*;
-    use crate::utils::proofs;
-    use ecdsa::to_eth_signed_message_hash;
-    use sp_core::{keccak_256, H160};
-    use sp_std::vec;
+	use ecdsa::to_eth_signed_message_hash;
+	use sp_core::{keccak_256, H160};
+	use sp_std::vec;
 
-    #[test]
-    fn proof_encode_decode() {
-        // Input params
-        let operators = vec![[1u8; 20], [2u8; 20]];
-        let weights = vec![100, 200];
-        let threshold = 99;
-        let signatures = vec![vec![1], vec![2]];
+	use super::*;
+	use crate::utils::proofs;
 
-        // Encode
-        let encoded = proofs::encode(
-            operators.clone(),
-            weights.clone(),
-            threshold,
-            signatures.clone(),
-        );
-        // Now decode
-        let proof = decode(&encoded).expect("Should decode proof");
+	#[test]
+	fn proof_encode_decode() {
+		// Input params
+		let operators = vec![[1u8; 20], [2u8; 20]];
+		let weights = vec![100, 200];
+		let threshold = 99;
+		let signatures = vec![vec![1], vec![2]];
 
-        let expected_proof = Proof {
-            operators: operators.into_iter().map(|x| Address::from(x)).collect(),
-            weights,
-            threshold,
-            signatures,
-        };
+		// Encode
+		let encoded = proofs::encode(
+			operators.clone(),
+			weights.clone(),
+			threshold,
+			signatures.clone(),
+		);
+		// Now decode
+		let proof = decode(&encoded).expect("Should decode proof");
 
-        assert_eq!(proof, expected_proof);
-    }
+		let expected_proof = Proof {
+			operators: operators.into_iter().map(|x| Address::from(x)).collect(),
+			weights,
+			threshold,
+			signatures,
+		};
 
-    #[test]
-    fn test_validate_signatures() {
-        let input = hex::decode("00000000000000000000000000000000000000000000000000000000000000400000000000000000000000000000000000000000000000000000000000001a4000000000000000000000000000000000000000000000000000000000000019e00000000000000000000000000000000000000000000000000000000000000089000000000000000000000000000000000000000000000000000000000000008000000000000000000000000000000000000000000000000000000000000003200000000000000000000000000000000000000000000000000000000000000ac00000000000000000000000000000000000000000000000000000000000000014024bd9868ae3a24be4618c06e99c4b8c8203c3ed5a7624926c7c1b2d5d4f2f830b61d8ae73e1fa78be41dca6e9eddaf58455c1fede28d9b659db46c2852199641f1bd2329cff9e675a7cfb713c3389e1cce070a52815dfe181e7bee84cebdf1632d682f903de5becca1d53ae70e10e89f9bbb08e34a44659c96cec5400c50aad3af28d1cd27a32b62d17b5b4f1f5dda694c038816ac210cd6041e69b307f68653d3102ef7fab39c68cf4877239f7fecf2390635ccd62188837cbaa92096f6a9f4258411491a045f5712f44d69d579a61fb32fc0e9e0a752da922722036807bdf4a471b6425920feaac5ea7e89a8184f31a8e07bfa7e2b3d8d13afe38c6f002075eed8a192323fa2a6426a9a76844c8795235c257122914b59460b7b93307a8085f25615b4a884455bdd7c3e15ab53bb7ce52c2da7de4c3f0271ed0e501d6b46b66bf20bc5dd1291380cd1968228d6f741ce65f11d65566440a14813dd29a1c45712848d49c6804fd0ee931a63d0aaffe723503c20154cf0ccbf40f4ead91ea247563e35f83efd31cd7ad9e8207b49ad3cac90e3ec0a3bfffbac92d54642b2d1e79538b3ca6bdb613981b8fb3300c64b904f61faf69fa8379419f5a256c076b949158484c76b6a6b3b6ca812fd48a75ad6fba67f234f5600e129de90aeb769fe1932a2dc27048ac32152e068c5d25c5ca69751f8a16d44f4426246817f3e54744b145640a7b52b2cd935d98c0358e1133af7d819ff48b65a5de5b461531574203c3a2c6ece32b9ff28b6004c21544f1235bb311dd00aa857dd1420fcfc0360eb1e055d1f67fde0977937822f0f553b812e92e4f5fe266e3c7dbd188e3ebbc8683f5e534eb6a5772039075f58b3c589a37d377783626a0a08ffbd822d19ff672510000000000000000000000000000000000000000000000000000000000000014000000000000000000000000000000000000000000000000000000000000028000000000000000000000000000000000000000000000000000000000000002c000000000000000000000000000000000000000000000000000000000000003000000000000000000000000000000000000000000000000000000000000000340000000000000000000000000000000000000000000000000000000000000038000000000000000000000000000000000000000000000000000000000000003c000000000000000000000000000000000000000000000000000000000000004000000000000000000000000000000000000000000000000000000000000000440000000000000000000000000000000000000000000000000000000000000048000000000000000000000000000000000000000000000000000000000000004c000000000000000000000000000000000000000000000000000000000000005000000000000000000000000000000000000000000000000000000000000000540000000000000000000000000000000000000000000000000000000000000058000000000000000000000000000000000000000000000000000000000000005c000000000000000000000000000000000000000000000000000000000000006000000000000000000000000000000000000000000000000000000000000000640000000000000000000000000000000000000000000000000000000000000068000000000000000000000000000000000000000000000000000000000000006c00000000000000000000000000000000000000000000000000000000000000700000000000000000000000000000000000000000000000000000000000000074000000000000000000000000000000000000000000000000000000000000000096275726e546f6b656e000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000096275726e546f6b656e000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000096275726e546f6b656e000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000096275726e546f6b656e000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000096275726e546f6b656e000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000096275726e546f6b656e000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000096275726e546f6b656e000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000096275726e546f6b656e000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000096275726e546f6b656e000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000096275726e546f6b656e000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000096275726e546f6b656e000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000096275726e546f6b656e000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000096275726e546f6b656e000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000096275726e546f6b656e000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000096275726e546f6b656e000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000096275726e546f6b656e000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000096275726e546f6b656e000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000096275726e546f6b656e000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000096275726e546f6b656e000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000096275726e546f6b656e000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000140000000000000000000000000000000000000000000000000000000000000280000000000000000000000000000000000000000000000000000000000000032000000000000000000000000000000000000000000000000000000000000003c00000000000000000000000000000000000000000000000000000000000000460000000000000000000000000000000000000000000000000000000000000050000000000000000000000000000000000000000000000000000000000000005a0000000000000000000000000000000000000000000000000000000000000064000000000000000000000000000000000000000000000000000000000000006e00000000000000000000000000000000000000000000000000000000000000780000000000000000000000000000000000000000000000000000000000000082000000000000000000000000000000000000000000000000000000000000008c000000000000000000000000000000000000000000000000000000000000009600000000000000000000000000000000000000000000000000000000000000a000000000000000000000000000000000000000000000000000000000000000aa00000000000000000000000000000000000000000000000000000000000000b400000000000000000000000000000000000000000000000000000000000000be00000000000000000000000000000000000000000000000000000000000000c800000000000000000000000000000000000000000000000000000000000000d200000000000000000000000000000000000000000000000000000000000000dc00000000000000000000000000000000000000000000000000000000000000e60000000000000000000000000000000000000000000000000000000000000008000000000000000000000000000000000000000000000000000000000000000400a6d1770549b534347933ca2f6f45745e6d5b479e7925fe84e63538bdfa8bbb2000000000000000000000000000000000000000000000000000000000000000341584c00000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000080000000000000000000000000000000000000000000000000000000000000004064f0cf9822dc7d35f96e1b5a8d710b89c56c781993bb8c0b121cec9d0b0e433a000000000000000000000000000000000000000000000000000000000000000341584c000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000800000000000000000000000000000000000000000000000000000000000000040f2b3576aee3d5ef0c8fa5bab2d67744de2b7f5f12e044ae4c5815f59c1c32637000000000000000000000000000000000000000000000000000000000000000761786c555344430000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000800000000000000000000000000000000000000000000000000000000000000040b4adc7f4169073d4b9acb8e333af363a5c1908198b31e7118fb712566583b784000000000000000000000000000000000000000000000000000000000000000341584c000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000800000000000000000000000000000000000000000000000000000000000000040ca2827d2512e5fc79794d5a9d1c796ad96286c1910187ad96e7ecf1ec5229d11000000000000000000000000000000000000000000000000000000000000000341584c0000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000008000000000000000000000000000000000000000000000000000000000000000407df01a951c4c9ca80588e361a4e16d8da8d86c1027a83ac362e88bdbc9d39d0c000000000000000000000000000000000000000000000000000000000000000341584c00000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000080000000000000000000000000000000000000000000000000000000000000004031b2e37c24f57b9f6c3e6293a2a9533a1aea2ab985a2736fdf408fe96e6a4190000000000000000000000000000000000000000000000000000000000000000341584c000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000800000000000000000000000000000000000000000000000000000000000000040fd9da3a2d08ea314fdce35b3679584ae7d0cf7fecc8598d5e5d6d32cdd17ff35000000000000000000000000000000000000000000000000000000000000000761786c555344430000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000800000000000000000000000000000000000000000000000000000000000000040a5b9348079667e180c95852ba7641fe3603b2446d0904dbcb11e08273ff6674d000000000000000000000000000000000000000000000000000000000000000761786c55534443000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000080000000000000000000000000000000000000000000000000000000000000004053d466178493c68d7df38e30e291dd9f7837d25ddef9eb126b23610d75eb8b3e000000000000000000000000000000000000000000000000000000000000000341584c000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000800000000000000000000000000000000000000000000000000000000000000040d4f3de95c9dc8f66795edc5219609bb874371e6d7fe688fbe15a92bc0ee0a889000000000000000000000000000000000000000000000000000000000000000341584c00000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000080000000000000000000000000000000000000000000000000000000000000004066b358b85bb956f4d44b79ddd2092e7e14b76ba91f4454b7665a9a6cc3504810000000000000000000000000000000000000000000000000000000000000000761786c5553444300000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000008000000000000000000000000000000000000000000000000000000000000000408c5ad2ffaa8ac0a1c403f589cb8acc726978dc85ef04431316676657611325a3000000000000000000000000000000000000000000000000000000000000000761786c555344430000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000800000000000000000000000000000000000000000000000000000000000000040dbead87253bba7961d78a8dc70d6f93bad72a5cc3808710aeb2e0b7a1f48136d000000000000000000000000000000000000000000000000000000000000000341584c000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000800000000000000000000000000000000000000000000000000000000000000040175a1d7d0687d58eba549b920c58f72acd85b3f32974bf715444d05a3fe5a7a7000000000000000000000000000000000000000000000000000000000000000341584c000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000800000000000000000000000000000000000000000000000000000000000000040b7fbff8df422b68c0e4260f635da3aa5d25e4f28c7bc55cdcd7d2ee05696c48f000000000000000000000000000000000000000000000000000000000000000341584c000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000800000000000000000000000000000000000000000000000000000000000000040fdc954879a4f2ec0f9d1f848a1e9bc16e554a89b3d07e09c7f1a5033ea9955c9000000000000000000000000000000000000000000000000000000000000000341584c0000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000008000000000000000000000000000000000000000000000000000000000000000400891068bb957a7ea53b723a07a56c6c58af9629c5074705e51dffafc7931100d000000000000000000000000000000000000000000000000000000000000000341584c00000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000080000000000000000000000000000000000000000000000000000000000000004033fcc5e8d848789577f23f4e23a9629f9d6a2aa0ad1d1b5e4a61fc860b0933fb000000000000000000000000000000000000000000000000000000000000000341584c000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000800000000000000000000000000000000000000000000000000000000000000040ddd2c66bd3fa3ee3ae6786a502ae3db87eb5d20c53534c0ac6c2c480c7f79581000000000000000000000000000000000000000000000000000000000000000341584c00000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001fe00000000000000000000000000000000000000000000000000000000000000080000000000000000000000000000000000000000000000000000000000000080000000000000000000000000000000000000000000000000000000000000178400000000000000000000000000000000000000000000000000000000000000f80000000000000000000000000000000000000000000000000000000000000003b00000000000000000000000006cef1a89423a4bd1ab859588e7c95e33e5dd5030000000000000000000000002416b62fae3cc9aaaa3d83560faaa099d65b9c6200000000000000000000000025b9437e280fd897e7b92fb969695a9ecfc679a600000000000000000000000025d57ab822c2d16ba0c26b2605204008bb2d5b760000000000000000000000002ab9393d3c4d39fe246f36740e349343aa45926e0000000000000000000000002e0be57ed7d7d0015e15495f5b2cffd1415ca9f40000000000000000000000002f885750185e9b0b50299f67619ccc5029b4fe780000000000000000000000003025973fbc55b3db307b22f5cf950f6e721e1e590000000000000000000000004cfa354d1391366409d8e15cfdaacb548166045300000000000000000000000052ab14b8f00bde612fd3a0efedb874a392e912a100000000000000000000000056a82852cf6fbe733d3bebd203345c8a3af4625f0000000000000000000000005c267d29094484dc6148a314f1da01476e81e3790000000000000000000000005d1bb3c6252377b0fee9a2f5784211e3bf6be9eb0000000000000000000000005f0d2b668452b194da9c4bb44039b46838a7fb4d0000000000000000000000005f709ffda63add6d93f9d5b0d6c8a93687f2060500000000000000000000000063897e57716a3136b367e6b04bdd148bb50743280000000000000000000000006c819a78f2f50957257f3f939533b355b9f53bc10000000000000000000000006c8e5111f0d7fa75f2bbff2c94e3743f4f509c890000000000000000000000006dc62faf15b9c63a30d60e8c51221628cd73441d0000000000000000000000006e9814e6d369d893d178ba086b79cab14456422b000000000000000000000000713643c8fdaf910c92caa2879133d3cf80e47cb7000000000000000000000000766075716f592113e3d89195670c3662b05b71a40000000000000000000000007a05ca9c569695b133c2c3f4363a2f05b87ae40b0000000000000000000000007cd4d2a644f9e480ea5180d8596e98f3bdf4e13900000000000000000000000083166fe68653e9431b36848d59bfaf17aceaed3c0000000000000000000000008b44ba5eb8e346dfe00a413c9d915b07a12b5492000000000000000000000000938e2dd4c5b0082c33d240c551ef7cc4b4bd89cc00000000000000000000000094e610ed1204199b69432a86e5ce2bc09b7b5c8d000000000000000000000000960cb2d7dd310d8f1011fa5e43efbac609e12f6a00000000000000000000000096c71ad79e004d547a6f4a7ca7ceb00e098a75db00000000000000000000000097dc92c1047638ce6cfb1e48680217cbf61605070000000000000000000000009c09e7e02e62d907565a6f8bd9a3f7f9519bb9a60000000000000000000000009ea41855d3caa93e41a7d09dff2161acad20a37f0000000000000000000000009eab40758dd776e32a3beb52c3eadf2d9dc8f59a000000000000000000000000a2f7ec24bf2e5b283d5ede4a70f2010353f43992000000000000000000000000a349424aeb5ebabd27e5ef0e948d1cf4e5bb5c93000000000000000000000000a3f5423cb27974aaa0c3f6626a14d7632ea47e2f000000000000000000000000a8be6d3f1ffa56a41a5e8328ab6684a0a8f3f797000000000000000000000000bb8302cc61bbca73cb5f1e33700b184dd27df39f000000000000000000000000bc550c0a985b8ba108425b81282d240d5efe025b000000000000000000000000be21095401b21b6f96a2f62368f38e374e3cdd4d000000000000000000000000bf1c1bf1135a9a8254af6d54df62acffe403df7d000000000000000000000000c00121d03080a59523397e495453a1bae10ac2ee000000000000000000000000caaacc0c574ed10f513cf4abee0878b80c5b2357000000000000000000000000ce283f39132fd4975d8fc3253f9979c9b614c458000000000000000000000000ce860111db05bd9d62099de889d53e998bbf212d000000000000000000000000d2031ae34371610669a5552e77eb83b2f72401d9000000000000000000000000d5f8c62eba364c3268462648dc1780c65a62f798000000000000000000000000e3a3b4dc387a86ae7b414c1931e17b2cda491b65000000000000000000000000e40ca7c2ddf7d30bdb504bf44e2f7bea087ccb92000000000000000000000000e4b2e50e5a780a3b28029f860101e0d0c64dc448000000000000000000000000e589143e35797e8a6f94d8be3c4dda6e7bc08cba000000000000000000000000e630337486f75373496acb6c25b3990c686637fd000000000000000000000000e7d3711884b278ea71c04ba478b669b0bc1f4501000000000000000000000000eb44d60895554f5bf4bff9734caa648ebf2598ce000000000000000000000000ef40aaa580dd27270cda1a2f58ab133b14d1aa6a000000000000000000000000f21d62e811aa647f8f8bec34acaf48a25a1e8b9a000000000000000000000000f8c8d60dca2de390fe4eda46cf801a46ee6cc529000000000000000000000000fc6f79efe7e3dbb8b63f74c8bf6ffc0225e11bca000000000000000000000000000000000000000000000000000000000000003b000000000000000000000000000000000000000000000000000000000000108d00000000000000000000000000000000000000000000000000000000000008fe0000000000000000000000000000000000000000000000000000000000000c2f0000000000000000000000000000000000000000000000000000000000000d9800000000000000000000000000000000000000000000000000000000000006de0000000000000000000000000000000000000000000000000000000000000e9600000000000000000000000000000000000000000000000000000000000008f10000000000000000000000000000000000000000000000000000000000000c030000000000000000000000000000000000000000000000000000000000000c8e00000000000000000000000000000000000000000000000000000000000002230000000000000000000000000000000000000000000000000000000000000891000000000000000000000000000000000000000000000000000000000000078300000000000000000000000000000000000000000000000000000000000002360000000000000000000000000000000000000000000000000000000000000cd90000000000000000000000000000000000000000000000000000000000000a140000000000000000000000000000000000000000000000000000000000000b0c00000000000000000000000000000000000000000000000000000000000002430000000000000000000000000000000000000000000000000000000000000224000000000000000000000000000000000000000000000000000000000000082600000000000000000000000000000000000000000000000000000000000001d90000000000000000000000000000000000000000000000000000000000000d08000000000000000000000000000000000000000000000000000000000000177500000000000000000000000000000000000000000000000000000000000019ff0000000000000000000000000000000000000000000000000000000000000a4f0000000000000000000000000000000000000000000000000000000000000978000000000000000000000000000000000000000000000000000000000000020f0000000000000000000000000000000000000000000000000000000000000baa0000000000000000000000000000000000000000000000000000000000000bef00000000000000000000000000000000000000000000000000000000000008ad000000000000000000000000000000000000000000000000000000000000058900000000000000000000000000000000000000000000000000000000000009b80000000000000000000000000000000000000000000000000000000000000c650000000000000000000000000000000000000000000000000000000000000dac0000000000000000000000000000000000000000000000000000000000000bc4000000000000000000000000000000000000000000000000000000000000118500000000000000000000000000000000000000000000000000000000000007b1000000000000000000000000000000000000000000000000000000000000103f000000000000000000000000000000000000000000000000000000000000022a0000000000000000000000000000000000000000000000000000000000000227000000000000000000000000000000000000000000000000000000000000098c00000000000000000000000000000000000000000000000000000000000005860000000000000000000000000000000000000000000000000000000000000bf90000000000000000000000000000000000000000000000000000000000000fea00000000000000000000000000000000000000000000000000000000000004c80000000000000000000000000000000000000000000000000000000000000f990000000000000000000000000000000000000000000000000000000000000c8500000000000000000000000000000000000000000000000000000000000006c00000000000000000000000000000000000000000000000000000000000000c2c00000000000000000000000000000000000000000000000000000000000006f70000000000000000000000000000000000000000000000000000000000000a220000000000000000000000000000000000000000000000000000000000000f3a00000000000000000000000000000000000000000000000000000000000008710000000000000000000000000000000000000000000000000000000000000cbc000000000000000000000000000000000000000000000000000000000000098900000000000000000000000000000000000000000000000000000000000013fa0000000000000000000000000000000000000000000000000000000000000f3f0000000000000000000000000000000000000000000000000000000000000d8300000000000000000000000000000000000000000000000000000000000004cc0000000000000000000000000000000000000000000000000000000000000c6c000000000000000000000000000000000000000000000000000000000000001a000000000000000000000000000000000000000000000000000000000000034000000000000000000000000000000000000000000000000000000000000003c0000000000000000000000000000000000000000000000000000000000000044000000000000000000000000000000000000000000000000000000000000004c0000000000000000000000000000000000000000000000000000000000000054000000000000000000000000000000000000000000000000000000000000005c0000000000000000000000000000000000000000000000000000000000000064000000000000000000000000000000000000000000000000000000000000006c0000000000000000000000000000000000000000000000000000000000000074000000000000000000000000000000000000000000000000000000000000007c0000000000000000000000000000000000000000000000000000000000000084000000000000000000000000000000000000000000000000000000000000008c0000000000000000000000000000000000000000000000000000000000000094000000000000000000000000000000000000000000000000000000000000009c00000000000000000000000000000000000000000000000000000000000000a400000000000000000000000000000000000000000000000000000000000000ac00000000000000000000000000000000000000000000000000000000000000b400000000000000000000000000000000000000000000000000000000000000bc00000000000000000000000000000000000000000000000000000000000000c400000000000000000000000000000000000000000000000000000000000000cc00000000000000000000000000000000000000000000000000000000000000d400000000000000000000000000000000000000000000000000000000000000dc00000000000000000000000000000000000000000000000000000000000000e400000000000000000000000000000000000000000000000000000000000000ec00000000000000000000000000000000000000000000000000000000000000f400000000000000000000000000000000000000000000000000000000000000fc0000000000000000000000000000000000000000000000000000000000000004178b04a351c892e310f836ce84111f226e2b59612dfb12b7e4620546fc07e830c28febdc4e7e1e979f65c25fb559a583e07c263b1a91c44ab67866d65ef9612631b000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000041d96eb274116cf66570235cf174c45ac6ac5fc3484fa705f855b59aa7966b68141d7bc3faeaefec754546ca59d0d13e17a1cdba29068fd01c92df9f0b71b686411c000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000041995574eae7662e388ec38823b1b177b21af0b834775131a4877759b91c3f9b5900f62999a22042b3d4ccf35caf0c67326d6e5673aabeb9e876927430fbb0c9251c00000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000004163e8fb22ec49dda417d33578cb254d4731dcd5bb3951fea20f9d2b4daee7f82e27f1edd0b4ae6df97f03089bb988852d2aefe760cfd13a368ee876c41a65cbb31b0000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000416d8576ece58978dfb244bc86cf03676b937a7e249dae8ac402781f92211a756c46883fd3e981cff438fa1c0ca367bb65070d085f2fdd620e2567b9d8d2c9dd831b000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000041bc2cf7de5a9d47955f16e861e86e63ccfbb80c94b261ecb8c7b49dce6b6d5add713c6873cddd9b5420d02e12cad6db949d351e23014508ba9d6aba63ae9391311c00000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000004158ce14379b885a73234072b602109a58f00b094bc30ce224d3316322a8cce7167bb821cd15fa5b000c03956e203ccbf59a47890f119af5037ab98eb1f40aa87a1b00000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000004150ab711dc34a02af44cc5497ae054574e2efe643ac99323b1ef2b8294566c0e0449331e64fedb6f80b472ad31bf84150f1ded1a1a24e55c3bed8f25003536ea61c000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000041af76d69ba2fcf9007e7301d71c1efe55a2175da759b01664c3a188203e8fb8684a4db0987da4c933676da809687369e9dffa067f48137c21f99cf012cc820e791b0000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000418dbeface745d74cabbdf5863e397bc06d40f3f11ac18c501d8204a87ed2d182f16d3bc5bf0c3726cbb1940165c6cda1735105b9ec437ac888e7aa9036182eadc1b0000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000415b748001a6d8ca7eb7e90a109cef53bceb26444885f343d8299d51d391bf784c7a0cb2659c95b373fd42db2a7af890431cd615710af2d81cbc7df484f21e2cd11c000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000041249210b837fa86b9f391b4ef14e212c9e039035653452815a5f8acf5c48228a657f47d8de2cbb497b093b101b1261d52131913ebc633f15ced26f62ad66a02341b000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000041b71d9b7d991deb9c7bebd82cb595d0d056abe9288a3759278a9bdf95fb3adf0c34a946b9ac4ec07c0ba1f0af04904d2621a7369cbc9af1b63aac6c3f81b4b0a41b00000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000004129e52b2f5f2779d46e17e1a842cf4433984c5c15c65dbc734f4289ee08387033557bf3127149ed9333aca3606585b115b60f315b22507c7d4ee6d6f43241be8b1b0000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000417f2f6840dac370c2afa6ab6a7e268688b7cd993a722557f47eced29bda96b47b7fc5328bc54e960d0680ab63a9ec38c699e12e45778809ccb1b206ee1446c5f41c000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000041577e94d9c6ce954f74cda9f1916bcabe6f79f5bdd13f8343d1a840df3e5cef663eef034e5807f8d8ada1ae657345fcf6cab721a8b457b85680261d6ac15ff0641b000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000041c71b153ec6b33ebbc4d85058cd7a5af7560326c198e874a4d9ebe3424444a5ec1e0adc7f17d309753c9e8c490b97be70b08a6a09bd647f4cf872e04bcf1198071c00000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000004102641dcf544f277e60f29f675f7332393c77e4ca0b6910713c9f63c46cf2cff20d5c5a621b6d96a3c809ef32af91552321836360255eda813d622026f931eefb1c00000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000004198f5b24330bc6c544f306de92a06c9f0a736c35ba4980d24ddef0889395288c81b21730e1c7c4d8b8ed10f167b46a1a1c0f8b9b8e032c0c29e940e6a3fd743011c000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000041e64ec926886ad800bc3a515b1ec587f918d45132e4117371f2c04f44d3cee0340d06b8a7719d0907b4b8add86165fd9544c81cb3a313bae51c4e486f765c88f31b000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000041919de19252b914c5a24e4e17a15334b55e3516736d77f9e3bbdd9eb2ed3772f95cf64bbebc5700d340df29e192b06c62cb9210bdb2a02f2a241f91b04d4b85e11c000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000041f36a88acf9563c11dd6b7260f1ae97875da0f283426d4c6d18af3e8b494c2d33153db701cf9e2dee49554d8ffeb6e1e6668870ab3d7f52b476d9f9dbeab0557f1b000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000041904b8ed91d65f227284331905b7c27c1c444d07cf117176c800a56403ce3bf717ebf9f9f22bda72e15f8927c2de0b588215ddeb843550b5d1ed64de30170273e1b000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000041bee456d591a1f9e09e445226d6ee80e929ae1f1e6938457d2073461b7389171360277233d20eb10a2a8e53645dd8a1aa95bb4a103d0d66276e4c1e51f36ed1f11b00000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000004127ae576319940182ab062e4342bd818e6fe4e7937cb81eb37a748f251fd04943484c58bf06dcc36eff3f12f01f6b29a2a6fb8a05aca2a67f501416b8b30ae18f1b000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000041c33ff523371e70dee70e0dfef77abfd1702f13c2998b95456f01bdcaf4dd4ebc31708efdad83ef022f52c00dbc9452ad67ed41e2f6253536b5035ce3572222891c00000000000000000000000000000000000000000000000000000000000000").expect("bad input");
-        let (msg, raw_proof) = proofs::decode_input(&input).expect("Input should decode");
-        let proof = decode(&raw_proof).expect("Proof should decode");
+		assert_eq!(proof, expected_proof);
+	}
 
-        let msg_hash = H256::from_slice(&to_eth_signed_message_hash(keccak_256(msg.as_slice())));
+	#[test]
+	fn test_validate_signatures() {
+		let input = hex::decode("00000000000000000000000000000000000000000000000000000000000000400000000000000000000000000000000000000000000000000000000000001a4000000000000000000000000000000000000000000000000000000000000019e00000000000000000000000000000000000000000000000000000000000000089000000000000000000000000000000000000000000000000000000000000008000000000000000000000000000000000000000000000000000000000000003200000000000000000000000000000000000000000000000000000000000000ac00000000000000000000000000000000000000000000000000000000000000014024bd9868ae3a24be4618c06e99c4b8c8203c3ed5a7624926c7c1b2d5d4f2f830b61d8ae73e1fa78be41dca6e9eddaf58455c1fede28d9b659db46c2852199641f1bd2329cff9e675a7cfb713c3389e1cce070a52815dfe181e7bee84cebdf1632d682f903de5becca1d53ae70e10e89f9bbb08e34a44659c96cec5400c50aad3af28d1cd27a32b62d17b5b4f1f5dda694c038816ac210cd6041e69b307f68653d3102ef7fab39c68cf4877239f7fecf2390635ccd62188837cbaa92096f6a9f4258411491a045f5712f44d69d579a61fb32fc0e9e0a752da922722036807bdf4a471b6425920feaac5ea7e89a8184f31a8e07bfa7e2b3d8d13afe38c6f002075eed8a192323fa2a6426a9a76844c8795235c257122914b59460b7b93307a8085f25615b4a884455bdd7c3e15ab53bb7ce52c2da7de4c3f0271ed0e501d6b46b66bf20bc5dd1291380cd1968228d6f741ce65f11d65566440a14813dd29a1c45712848d49c6804fd0ee931a63d0aaffe723503c20154cf0ccbf40f4ead91ea247563e35f83efd31cd7ad9e8207b49ad3cac90e3ec0a3bfffbac92d54642b2d1e79538b3ca6bdb613981b8fb3300c64b904f61faf69fa8379419f5a256c076b949158484c76b6a6b3b6ca812fd48a75ad6fba67f234f5600e129de90aeb769fe1932a2dc27048ac32152e068c5d25c5ca69751f8a16d44f4426246817f3e54744b145640a7b52b2cd935d98c0358e1133af7d819ff48b65a5de5b461531574203c3a2c6ece32b9ff28b6004c21544f1235bb311dd00aa857dd1420fcfc0360eb1e055d1f67fde0977937822f0f553b812e92e4f5fe266e3c7dbd188e3ebbc8683f5e534eb6a5772039075f58b3c589a37d377783626a0a08ffbd822d19ff672510000000000000000000000000000000000000000000000000000000000000014000000000000000000000000000000000000000000000000000000000000028000000000000000000000000000000000000000000000000000000000000002c000000000000000000000000000000000000000000000000000000000000003000000000000000000000000000000000000000000000000000000000000000340000000000000000000000000000000000000000000000000000000000000038000000000000000000000000000000000000000000000000000000000000003c000000000000000000000000000000000000000000000000000000000000004000000000000000000000000000000000000000000000000000000000000000440000000000000000000000000000000000000000000000000000000000000048000000000000000000000000000000000000000000000000000000000000004c000000000000000000000000000000000000000000000000000000000000005000000000000000000000000000000000000000000000000000000000000000540000000000000000000000000000000000000000000000000000000000000058000000000000000000000000000000000000000000000000000000000000005c000000000000000000000000000000000000000000000000000000000000006000000000000000000000000000000000000000000000000000000000000000640000000000000000000000000000000000000000000000000000000000000068000000000000000000000000000000000000000000000000000000000000006c00000000000000000000000000000000000000000000000000000000000000700000000000000000000000000000000000000000000000000000000000000074000000000000000000000000000000000000000000000000000000000000000096275726e546f6b656e000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000096275726e546f6b656e000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000096275726e546f6b656e000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000096275726e546f6b656e000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000096275726e546f6b656e000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000096275726e546f6b656e000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000096275726e546f6b656e000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000096275726e546f6b656e000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000096275726e546f6b656e000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000096275726e546f6b656e000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000096275726e546f6b656e000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000096275726e546f6b656e000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000096275726e546f6b656e000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000096275726e546f6b656e000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000096275726e546f6b656e000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000096275726e546f6b656e000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000096275726e546f6b656e000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000096275726e546f6b656e000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000096275726e546f6b656e000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000096275726e546f6b656e000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000140000000000000000000000000000000000000000000000000000000000000280000000000000000000000000000000000000000000000000000000000000032000000000000000000000000000000000000000000000000000000000000003c00000000000000000000000000000000000000000000000000000000000000460000000000000000000000000000000000000000000000000000000000000050000000000000000000000000000000000000000000000000000000000000005a0000000000000000000000000000000000000000000000000000000000000064000000000000000000000000000000000000000000000000000000000000006e00000000000000000000000000000000000000000000000000000000000000780000000000000000000000000000000000000000000000000000000000000082000000000000000000000000000000000000000000000000000000000000008c000000000000000000000000000000000000000000000000000000000000009600000000000000000000000000000000000000000000000000000000000000a000000000000000000000000000000000000000000000000000000000000000aa00000000000000000000000000000000000000000000000000000000000000b400000000000000000000000000000000000000000000000000000000000000be00000000000000000000000000000000000000000000000000000000000000c800000000000000000000000000000000000000000000000000000000000000d200000000000000000000000000000000000000000000000000000000000000dc00000000000000000000000000000000000000000000000000000000000000e60000000000000000000000000000000000000000000000000000000000000008000000000000000000000000000000000000000000000000000000000000000400a6d1770549b534347933ca2f6f45745e6d5b479e7925fe84e63538bdfa8bbb2000000000000000000000000000000000000000000000000000000000000000341584c00000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000080000000000000000000000000000000000000000000000000000000000000004064f0cf9822dc7d35f96e1b5a8d710b89c56c781993bb8c0b121cec9d0b0e433a000000000000000000000000000000000000000000000000000000000000000341584c000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000800000000000000000000000000000000000000000000000000000000000000040f2b3576aee3d5ef0c8fa5bab2d67744de2b7f5f12e044ae4c5815f59c1c32637000000000000000000000000000000000000000000000000000000000000000761786c555344430000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000800000000000000000000000000000000000000000000000000000000000000040b4adc7f4169073d4b9acb8e333af363a5c1908198b31e7118fb712566583b784000000000000000000000000000000000000000000000000000000000000000341584c000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000800000000000000000000000000000000000000000000000000000000000000040ca2827d2512e5fc79794d5a9d1c796ad96286c1910187ad96e7ecf1ec5229d11000000000000000000000000000000000000000000000000000000000000000341584c0000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000008000000000000000000000000000000000000000000000000000000000000000407df01a951c4c9ca80588e361a4e16d8da8d86c1027a83ac362e88bdbc9d39d0c000000000000000000000000000000000000000000000000000000000000000341584c00000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000080000000000000000000000000000000000000000000000000000000000000004031b2e37c24f57b9f6c3e6293a2a9533a1aea2ab985a2736fdf408fe96e6a4190000000000000000000000000000000000000000000000000000000000000000341584c000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000800000000000000000000000000000000000000000000000000000000000000040fd9da3a2d08ea314fdce35b3679584ae7d0cf7fecc8598d5e5d6d32cdd17ff35000000000000000000000000000000000000000000000000000000000000000761786c555344430000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000800000000000000000000000000000000000000000000000000000000000000040a5b9348079667e180c95852ba7641fe3603b2446d0904dbcb11e08273ff6674d000000000000000000000000000000000000000000000000000000000000000761786c55534443000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000080000000000000000000000000000000000000000000000000000000000000004053d466178493c68d7df38e30e291dd9f7837d25ddef9eb126b23610d75eb8b3e000000000000000000000000000000000000000000000000000000000000000341584c000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000800000000000000000000000000000000000000000000000000000000000000040d4f3de95c9dc8f66795edc5219609bb874371e6d7fe688fbe15a92bc0ee0a889000000000000000000000000000000000000000000000000000000000000000341584c00000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000080000000000000000000000000000000000000000000000000000000000000004066b358b85bb956f4d44b79ddd2092e7e14b76ba91f4454b7665a9a6cc3504810000000000000000000000000000000000000000000000000000000000000000761786c5553444300000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000008000000000000000000000000000000000000000000000000000000000000000408c5ad2ffaa8ac0a1c403f589cb8acc726978dc85ef04431316676657611325a3000000000000000000000000000000000000000000000000000000000000000761786c555344430000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000800000000000000000000000000000000000000000000000000000000000000040dbead87253bba7961d78a8dc70d6f93bad72a5cc3808710aeb2e0b7a1f48136d000000000000000000000000000000000000000000000000000000000000000341584c000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000800000000000000000000000000000000000000000000000000000000000000040175a1d7d0687d58eba549b920c58f72acd85b3f32974bf715444d05a3fe5a7a7000000000000000000000000000000000000000000000000000000000000000341584c000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000800000000000000000000000000000000000000000000000000000000000000040b7fbff8df422b68c0e4260f635da3aa5d25e4f28c7bc55cdcd7d2ee05696c48f000000000000000000000000000000000000000000000000000000000000000341584c000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000800000000000000000000000000000000000000000000000000000000000000040fdc954879a4f2ec0f9d1f848a1e9bc16e554a89b3d07e09c7f1a5033ea9955c9000000000000000000000000000000000000000000000000000000000000000341584c0000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000008000000000000000000000000000000000000000000000000000000000000000400891068bb957a7ea53b723a07a56c6c58af9629c5074705e51dffafc7931100d000000000000000000000000000000000000000000000000000000000000000341584c00000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000080000000000000000000000000000000000000000000000000000000000000004033fcc5e8d848789577f23f4e23a9629f9d6a2aa0ad1d1b5e4a61fc860b0933fb000000000000000000000000000000000000000000000000000000000000000341584c000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000800000000000000000000000000000000000000000000000000000000000000040ddd2c66bd3fa3ee3ae6786a502ae3db87eb5d20c53534c0ac6c2c480c7f79581000000000000000000000000000000000000000000000000000000000000000341584c00000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001fe00000000000000000000000000000000000000000000000000000000000000080000000000000000000000000000000000000000000000000000000000000080000000000000000000000000000000000000000000000000000000000000178400000000000000000000000000000000000000000000000000000000000000f80000000000000000000000000000000000000000000000000000000000000003b00000000000000000000000006cef1a89423a4bd1ab859588e7c95e33e5dd5030000000000000000000000002416b62fae3cc9aaaa3d83560faaa099d65b9c6200000000000000000000000025b9437e280fd897e7b92fb969695a9ecfc679a600000000000000000000000025d57ab822c2d16ba0c26b2605204008bb2d5b760000000000000000000000002ab9393d3c4d39fe246f36740e349343aa45926e0000000000000000000000002e0be57ed7d7d0015e15495f5b2cffd1415ca9f40000000000000000000000002f885750185e9b0b50299f67619ccc5029b4fe780000000000000000000000003025973fbc55b3db307b22f5cf950f6e721e1e590000000000000000000000004cfa354d1391366409d8e15cfdaacb548166045300000000000000000000000052ab14b8f00bde612fd3a0efedb874a392e912a100000000000000000000000056a82852cf6fbe733d3bebd203345c8a3af4625f0000000000000000000000005c267d29094484dc6148a314f1da01476e81e3790000000000000000000000005d1bb3c6252377b0fee9a2f5784211e3bf6be9eb0000000000000000000000005f0d2b668452b194da9c4bb44039b46838a7fb4d0000000000000000000000005f709ffda63add6d93f9d5b0d6c8a93687f2060500000000000000000000000063897e57716a3136b367e6b04bdd148bb50743280000000000000000000000006c819a78f2f50957257f3f939533b355b9f53bc10000000000000000000000006c8e5111f0d7fa75f2bbff2c94e3743f4f509c890000000000000000000000006dc62faf15b9c63a30d60e8c51221628cd73441d0000000000000000000000006e9814e6d369d893d178ba086b79cab14456422b000000000000000000000000713643c8fdaf910c92caa2879133d3cf80e47cb7000000000000000000000000766075716f592113e3d89195670c3662b05b71a40000000000000000000000007a05ca9c569695b133c2c3f4363a2f05b87ae40b0000000000000000000000007cd4d2a644f9e480ea5180d8596e98f3bdf4e13900000000000000000000000083166fe68653e9431b36848d59bfaf17aceaed3c0000000000000000000000008b44ba5eb8e346dfe00a413c9d915b07a12b5492000000000000000000000000938e2dd4c5b0082c33d240c551ef7cc4b4bd89cc00000000000000000000000094e610ed1204199b69432a86e5ce2bc09b7b5c8d000000000000000000000000960cb2d7dd310d8f1011fa5e43efbac609e12f6a00000000000000000000000096c71ad79e004d547a6f4a7ca7ceb00e098a75db00000000000000000000000097dc92c1047638ce6cfb1e48680217cbf61605070000000000000000000000009c09e7e02e62d907565a6f8bd9a3f7f9519bb9a60000000000000000000000009ea41855d3caa93e41a7d09dff2161acad20a37f0000000000000000000000009eab40758dd776e32a3beb52c3eadf2d9dc8f59a000000000000000000000000a2f7ec24bf2e5b283d5ede4a70f2010353f43992000000000000000000000000a349424aeb5ebabd27e5ef0e948d1cf4e5bb5c93000000000000000000000000a3f5423cb27974aaa0c3f6626a14d7632ea47e2f000000000000000000000000a8be6d3f1ffa56a41a5e8328ab6684a0a8f3f797000000000000000000000000bb8302cc61bbca73cb5f1e33700b184dd27df39f000000000000000000000000bc550c0a985b8ba108425b81282d240d5efe025b000000000000000000000000be21095401b21b6f96a2f62368f38e374e3cdd4d000000000000000000000000bf1c1bf1135a9a8254af6d54df62acffe403df7d000000000000000000000000c00121d03080a59523397e495453a1bae10ac2ee000000000000000000000000caaacc0c574ed10f513cf4abee0878b80c5b2357000000000000000000000000ce283f39132fd4975d8fc3253f9979c9b614c458000000000000000000000000ce860111db05bd9d62099de889d53e998bbf212d000000000000000000000000d2031ae34371610669a5552e77eb83b2f72401d9000000000000000000000000d5f8c62eba364c3268462648dc1780c65a62f798000000000000000000000000e3a3b4dc387a86ae7b414c1931e17b2cda491b65000000000000000000000000e40ca7c2ddf7d30bdb504bf44e2f7bea087ccb92000000000000000000000000e4b2e50e5a780a3b28029f860101e0d0c64dc448000000000000000000000000e589143e35797e8a6f94d8be3c4dda6e7bc08cba000000000000000000000000e630337486f75373496acb6c25b3990c686637fd000000000000000000000000e7d3711884b278ea71c04ba478b669b0bc1f4501000000000000000000000000eb44d60895554f5bf4bff9734caa648ebf2598ce000000000000000000000000ef40aaa580dd27270cda1a2f58ab133b14d1aa6a000000000000000000000000f21d62e811aa647f8f8bec34acaf48a25a1e8b9a000000000000000000000000f8c8d60dca2de390fe4eda46cf801a46ee6cc529000000000000000000000000fc6f79efe7e3dbb8b63f74c8bf6ffc0225e11bca000000000000000000000000000000000000000000000000000000000000003b000000000000000000000000000000000000000000000000000000000000108d00000000000000000000000000000000000000000000000000000000000008fe0000000000000000000000000000000000000000000000000000000000000c2f0000000000000000000000000000000000000000000000000000000000000d9800000000000000000000000000000000000000000000000000000000000006de0000000000000000000000000000000000000000000000000000000000000e9600000000000000000000000000000000000000000000000000000000000008f10000000000000000000000000000000000000000000000000000000000000c030000000000000000000000000000000000000000000000000000000000000c8e00000000000000000000000000000000000000000000000000000000000002230000000000000000000000000000000000000000000000000000000000000891000000000000000000000000000000000000000000000000000000000000078300000000000000000000000000000000000000000000000000000000000002360000000000000000000000000000000000000000000000000000000000000cd90000000000000000000000000000000000000000000000000000000000000a140000000000000000000000000000000000000000000000000000000000000b0c00000000000000000000000000000000000000000000000000000000000002430000000000000000000000000000000000000000000000000000000000000224000000000000000000000000000000000000000000000000000000000000082600000000000000000000000000000000000000000000000000000000000001d90000000000000000000000000000000000000000000000000000000000000d08000000000000000000000000000000000000000000000000000000000000177500000000000000000000000000000000000000000000000000000000000019ff0000000000000000000000000000000000000000000000000000000000000a4f0000000000000000000000000000000000000000000000000000000000000978000000000000000000000000000000000000000000000000000000000000020f0000000000000000000000000000000000000000000000000000000000000baa0000000000000000000000000000000000000000000000000000000000000bef00000000000000000000000000000000000000000000000000000000000008ad000000000000000000000000000000000000000000000000000000000000058900000000000000000000000000000000000000000000000000000000000009b80000000000000000000000000000000000000000000000000000000000000c650000000000000000000000000000000000000000000000000000000000000dac0000000000000000000000000000000000000000000000000000000000000bc4000000000000000000000000000000000000000000000000000000000000118500000000000000000000000000000000000000000000000000000000000007b1000000000000000000000000000000000000000000000000000000000000103f000000000000000000000000000000000000000000000000000000000000022a0000000000000000000000000000000000000000000000000000000000000227000000000000000000000000000000000000000000000000000000000000098c00000000000000000000000000000000000000000000000000000000000005860000000000000000000000000000000000000000000000000000000000000bf90000000000000000000000000000000000000000000000000000000000000fea00000000000000000000000000000000000000000000000000000000000004c80000000000000000000000000000000000000000000000000000000000000f990000000000000000000000000000000000000000000000000000000000000c8500000000000000000000000000000000000000000000000000000000000006c00000000000000000000000000000000000000000000000000000000000000c2c00000000000000000000000000000000000000000000000000000000000006f70000000000000000000000000000000000000000000000000000000000000a220000000000000000000000000000000000000000000000000000000000000f3a00000000000000000000000000000000000000000000000000000000000008710000000000000000000000000000000000000000000000000000000000000cbc000000000000000000000000000000000000000000000000000000000000098900000000000000000000000000000000000000000000000000000000000013fa0000000000000000000000000000000000000000000000000000000000000f3f0000000000000000000000000000000000000000000000000000000000000d8300000000000000000000000000000000000000000000000000000000000004cc0000000000000000000000000000000000000000000000000000000000000c6c000000000000000000000000000000000000000000000000000000000000001a000000000000000000000000000000000000000000000000000000000000034000000000000000000000000000000000000000000000000000000000000003c0000000000000000000000000000000000000000000000000000000000000044000000000000000000000000000000000000000000000000000000000000004c0000000000000000000000000000000000000000000000000000000000000054000000000000000000000000000000000000000000000000000000000000005c0000000000000000000000000000000000000000000000000000000000000064000000000000000000000000000000000000000000000000000000000000006c0000000000000000000000000000000000000000000000000000000000000074000000000000000000000000000000000000000000000000000000000000007c0000000000000000000000000000000000000000000000000000000000000084000000000000000000000000000000000000000000000000000000000000008c0000000000000000000000000000000000000000000000000000000000000094000000000000000000000000000000000000000000000000000000000000009c00000000000000000000000000000000000000000000000000000000000000a400000000000000000000000000000000000000000000000000000000000000ac00000000000000000000000000000000000000000000000000000000000000b400000000000000000000000000000000000000000000000000000000000000bc00000000000000000000000000000000000000000000000000000000000000c400000000000000000000000000000000000000000000000000000000000000cc00000000000000000000000000000000000000000000000000000000000000d400000000000000000000000000000000000000000000000000000000000000dc00000000000000000000000000000000000000000000000000000000000000e400000000000000000000000000000000000000000000000000000000000000ec00000000000000000000000000000000000000000000000000000000000000f400000000000000000000000000000000000000000000000000000000000000fc0000000000000000000000000000000000000000000000000000000000000004178b04a351c892e310f836ce84111f226e2b59612dfb12b7e4620546fc07e830c28febdc4e7e1e979f65c25fb559a583e07c263b1a91c44ab67866d65ef9612631b000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000041d96eb274116cf66570235cf174c45ac6ac5fc3484fa705f855b59aa7966b68141d7bc3faeaefec754546ca59d0d13e17a1cdba29068fd01c92df9f0b71b686411c000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000041995574eae7662e388ec38823b1b177b21af0b834775131a4877759b91c3f9b5900f62999a22042b3d4ccf35caf0c67326d6e5673aabeb9e876927430fbb0c9251c00000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000004163e8fb22ec49dda417d33578cb254d4731dcd5bb3951fea20f9d2b4daee7f82e27f1edd0b4ae6df97f03089bb988852d2aefe760cfd13a368ee876c41a65cbb31b0000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000416d8576ece58978dfb244bc86cf03676b937a7e249dae8ac402781f92211a756c46883fd3e981cff438fa1c0ca367bb65070d085f2fdd620e2567b9d8d2c9dd831b000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000041bc2cf7de5a9d47955f16e861e86e63ccfbb80c94b261ecb8c7b49dce6b6d5add713c6873cddd9b5420d02e12cad6db949d351e23014508ba9d6aba63ae9391311c00000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000004158ce14379b885a73234072b602109a58f00b094bc30ce224d3316322a8cce7167bb821cd15fa5b000c03956e203ccbf59a47890f119af5037ab98eb1f40aa87a1b00000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000004150ab711dc34a02af44cc5497ae054574e2efe643ac99323b1ef2b8294566c0e0449331e64fedb6f80b472ad31bf84150f1ded1a1a24e55c3bed8f25003536ea61c000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000041af76d69ba2fcf9007e7301d71c1efe55a2175da759b01664c3a188203e8fb8684a4db0987da4c933676da809687369e9dffa067f48137c21f99cf012cc820e791b0000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000418dbeface745d74cabbdf5863e397bc06d40f3f11ac18c501d8204a87ed2d182f16d3bc5bf0c3726cbb1940165c6cda1735105b9ec437ac888e7aa9036182eadc1b0000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000415b748001a6d8ca7eb7e90a109cef53bceb26444885f343d8299d51d391bf784c7a0cb2659c95b373fd42db2a7af890431cd615710af2d81cbc7df484f21e2cd11c000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000041249210b837fa86b9f391b4ef14e212c9e039035653452815a5f8acf5c48228a657f47d8de2cbb497b093b101b1261d52131913ebc633f15ced26f62ad66a02341b000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000041b71d9b7d991deb9c7bebd82cb595d0d056abe9288a3759278a9bdf95fb3adf0c34a946b9ac4ec07c0ba1f0af04904d2621a7369cbc9af1b63aac6c3f81b4b0a41b00000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000004129e52b2f5f2779d46e17e1a842cf4433984c5c15c65dbc734f4289ee08387033557bf3127149ed9333aca3606585b115b60f315b22507c7d4ee6d6f43241be8b1b0000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000417f2f6840dac370c2afa6ab6a7e268688b7cd993a722557f47eced29bda96b47b7fc5328bc54e960d0680ab63a9ec38c699e12e45778809ccb1b206ee1446c5f41c000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000041577e94d9c6ce954f74cda9f1916bcabe6f79f5bdd13f8343d1a840df3e5cef663eef034e5807f8d8ada1ae657345fcf6cab721a8b457b85680261d6ac15ff0641b000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000041c71b153ec6b33ebbc4d85058cd7a5af7560326c198e874a4d9ebe3424444a5ec1e0adc7f17d309753c9e8c490b97be70b08a6a09bd647f4cf872e04bcf1198071c00000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000004102641dcf544f277e60f29f675f7332393c77e4ca0b6910713c9f63c46cf2cff20d5c5a621b6d96a3c809ef32af91552321836360255eda813d622026f931eefb1c00000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000004198f5b24330bc6c544f306de92a06c9f0a736c35ba4980d24ddef0889395288c81b21730e1c7c4d8b8ed10f167b46a1a1c0f8b9b8e032c0c29e940e6a3fd743011c000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000041e64ec926886ad800bc3a515b1ec587f918d45132e4117371f2c04f44d3cee0340d06b8a7719d0907b4b8add86165fd9544c81cb3a313bae51c4e486f765c88f31b000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000041919de19252b914c5a24e4e17a15334b55e3516736d77f9e3bbdd9eb2ed3772f95cf64bbebc5700d340df29e192b06c62cb9210bdb2a02f2a241f91b04d4b85e11c000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000041f36a88acf9563c11dd6b7260f1ae97875da0f283426d4c6d18af3e8b494c2d33153db701cf9e2dee49554d8ffeb6e1e6668870ab3d7f52b476d9f9dbeab0557f1b000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000041904b8ed91d65f227284331905b7c27c1c444d07cf117176c800a56403ce3bf717ebf9f9f22bda72e15f8927c2de0b588215ddeb843550b5d1ed64de30170273e1b000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000041bee456d591a1f9e09e445226d6ee80e929ae1f1e6938457d2073461b7389171360277233d20eb10a2a8e53645dd8a1aa95bb4a103d0d66276e4c1e51f36ed1f11b00000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000004127ae576319940182ab062e4342bd818e6fe4e7937cb81eb37a748f251fd04943484c58bf06dcc36eff3f12f01f6b29a2a6fb8a05aca2a67f501416b8b30ae18f1b000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000041c33ff523371e70dee70e0dfef77abfd1702f13c2998b95456f01bdcaf4dd4ebc31708efdad83ef022f52c00dbc9452ad67ed41e2f6253536b5035ce3572222891c00000000000000000000000000000000000000000000000000000000000000").expect("bad input");
+		let (msg, raw_proof) = proofs::decode_input(&input).expect("Input should decode");
+		let proof = decode(&raw_proof).expect("Proof should decode");
 
-        assert!(validate_signatures(msg_hash, proof,).is_ok());
-    }
+		let msg_hash = H256::from_slice(&to_eth_signed_message_hash(keccak_256(msg.as_slice())));
 
-    #[test]
-    fn test_sign_verify_flow() {
-        let msg = vec![0_u8, 1_u8];
-        let msg_hash = H256::from_slice(&to_eth_signed_message_hash(keccak_256(msg.as_slice())));
-        let (public, secret) = ecdsa::generate_keypair();
-        let pk_hash = H160::from(H256::from_slice(keccak_256(&public).as_slice()));
-        let sig = ecdsa::sign_message(msg_hash, &secret);
-        assert_eq!(sig.len(), 65);
-        match ecdsa::recover(msg_hash, sig).ok() {
-            None => {
-                panic!("should not happen")
-            }
-            Some(signer) => {
-                assert_eq!(pk_hash.0, signer.0)
-            }
-        }
-    }
+		assert!(validate_signatures(msg_hash, proof,).is_ok());
+	}
+
+	#[test]
+	fn test_sign_verify_flow() {
+		let msg = vec![0_u8, 1_u8];
+		let msg_hash = H256::from_slice(&to_eth_signed_message_hash(keccak_256(msg.as_slice())));
+		let (public, secret) = ecdsa::generate_keypair();
+		let pk_hash = H160::from(H256::from_slice(keccak_256(&public).as_slice()));
+		let sig = ecdsa::sign_message(msg_hash, &secret);
+		assert_eq!(sig.len(), 65);
+		match ecdsa::recover(msg_hash, sig).ok() {
+			None => {
+				panic!("should not happen")
+			}
+			Some(signer) => {
+				assert_eq!(pk_hash.0, signer.0)
+			}
+		}
+	}
 }
