+++ integration-test/Cargo.toml
@@ -8,7 +8,6 @@ edition = "2021"
 [dependencies]
 serde = { version = "1.0.136", optional = true }
 codec = { package = "parity-scale-codec", version = "3.0.0", default-features = false }
-proc-macro2 = "1.0.40"
 hex = "0.4.3"
 ethabi = { version = "16.0", default-features = false }
 
@@ -31,7 +30,7 @@ kusama-runtime = { git = "https://github.com/paritytech/polkadot", branch = "rel
 pallet-xcm = { git = "https://github.com/paritytech/polkadot", branch = "release-v0.9.33" }
 
 xcm-emulator = { git = "https://github.com/shaunxw/xcm-simulator", rev = "c62dbf8e247c03d9797dbf3f1886f74973f2b22a" }
-sample-runtime = { path = "../sample-runtime" }
+sample-runtime = { path = "sample-runtime" }
 axelar-cgp = { path = "../gateway"}
-sample-receiver = { path = "../sample-receiver"}
+sample-receiver = { path = "sample-receiver" }
 ecdsa = { path = "../libs/ecdsa" }
