+++ integration-test/sample-receiver/Cargo.toml
@@ -11,15 +11,14 @@ edition = "2021"
 codec = { package = "parity-scale-codec", version = "3.0.0", default-features = false, features = [
   "max-encoded-len",
 ] }
-hex = "0.4.3"
 scale-info = { version = "2.1.2", default-features = false, features = [
   "derive",
 ] }
-serde = { version = "1.0.136", optional = true }
 ethabi = { version = "16.0", default-features = false }
 
 # Local
-ecdsa = { path = "../libs/ecdsa" }
+ecdsa = { path = "../../libs/ecdsa", default-features = false }
+hex = { version = "0.4.3", default-features = false, features = ["alloc"] }
 
 # Substrate
 frame-support = { git = "https://github.com/paritytech/substrate", default-features = false, branch = "polkadot-v0.9.33" }
@@ -38,27 +37,44 @@ polkadot-parachain = { git = "https://github.com/paritytech/polkadot", default-f
 cumulus-pallet-xcm = { git = "https://github.com/paritytech/cumulus", default-features = false, branch = "polkadot-v0.9.33" }
 cumulus-primitives-core = { git = "https://github.com/paritytech/cumulus", default-features = false , branch = "polkadot-v0.9.33" }
 
+# Benchmarking
+frame-benchmarking = { git = "https://github.com/paritytech/substrate", default-features = false, optional = true, branch = "polkadot-v0.9.33" }
+
 [dev-dependencies]
-xcm-executor = { git = "https://github.com/paritytech/polkadot", default-features = false, branch = "release-v0.9.33" }
-xcm-builder = { git = "https://github.com/paritytech/polkadot", default-features = false, branch = "release-v0.9.33" }
-parachain-info = { git = "https://github.com/paritytech/cumulus", default-features = false , branch = "polkadot-v0.9.33" }
+xcm-executor = { git = "https://github.com/paritytech/polkadot", branch = "release-v0.9.33" }
+xcm-builder = { git = "https://github.com/paritytech/polkadot", branch = "release-v0.9.33" }
+parachain-info = { git = "https://github.com/paritytech/cumulus", branch = "polkadot-v0.9.33" }
 
 
 [features]
 default = ['std']
 std = [
-  "serde",
-
+  "hex/std",
+  "scale-info/std",
+  "ethabi/std",
   "codec/std",
   "frame-support/std",
   "frame-system/std",
   "scale-info/std",
   "sp-runtime/std",
   "sp-std/std",
+  "sp-io/std",
+  "sp-core/std",
+  "xcm/std",
+  "pallet-xcm/std",
+  "polkadot-parachain/std",
+  "cumulus-pallet-xcm/std",
+  "cumulus-primitives-core/std",
+  "ecdsa/std"
 ]
 runtime-benchmarks = [
+  "frame-benchmarking",
   "frame-support/runtime-benchmarks",
   "frame-system/runtime-benchmarks",
   "sp-runtime/runtime-benchmarks",
+  "xcm/runtime-benchmarks",
+  "pallet-xcm/runtime-benchmarks",
+  "polkadot-parachain/runtime-benchmarks",
 ]
+
 try-runtime = ["frame-support/try-runtime", "frame-system/try-runtime"]
