+++ integration-test/src/lib.rs
@@ -1,363 +1,377 @@
+// Copyright 2023 Centrifuge Foundation (centrifuge.io).
+//
+// This file is part of the axelar-cgp-substrate project.
+// Axelar-cgp-substrate is free software: you can redistribute it and/or modify
+// it under the terms of the GNU General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version (see http://www.gnu.org/licenses).
+// Centrifuge is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU General Public License for more details.
+
 mod env;
 
 #[cfg(test)]
 mod tests {
-    use super::*;
-    use axelar_cgp::proof;
-    use codec::Encode;
-    use env::{Network, Para1, Para2};
-    use ethabi::Token;
-    use frame_support::{assert_noop, assert_ok};
-    use sample_runtime::{AccountId, RuntimeOrigin};
-    use sp_core::{keccak_256, H160, H256, U256};
-    use sp_runtime::DispatchError::BadOrigin;
-    use std::str::FromStr;
-    use xcm::latest::prelude::*;
-    use xcm_emulator::TestExt;
+	use std::str::FromStr;
 
-    #[test]
-    fn transact_xcm_origins() {
-        Network::reset();
+	use axelar_cgp::proof;
+	use codec::Encode;
+	use env::{Network, Para1, Para2};
+	use ethabi::Token;
+	use frame_support::{assert_noop, assert_ok};
+	use sample_runtime::{AccountId, RuntimeOrigin};
+	use sp_core::{keccak_256, H160, H256, U256};
+	use sp_runtime::DispatchError::BadOrigin;
+	use xcm::latest::prelude::*;
+	use xcm_emulator::TestExt;
 
-        Para1::execute_with(|| {
-            let inner_call =
-                sample_runtime::RuntimeCall::System(frame_system::Call::remark_with_event {
-                    remark: vec![10],
-                });
+	use super::*;
 
-            let fee_asset = MultiAsset {
-                id: Concrete(MultiLocation {
-                    parents: 1,
-                    interior: Junctions::Here,
-                }),
-                fun: Fungible(8_000_000_000),
-            };
+	#[test]
+	fn transact_xcm_origins() {
+		Network::reset();
 
-            assert_ok!(sample_runtime::PolkadotXcm::send_xcm(
-                Here,
-                MultiLocation::new(1, X1(Parachain(2))),
-                Xcm(vec![
-                    WithdrawAsset(fee_asset.clone().into()),
-                    BuyExecution {
-                        fees: fee_asset.into(),
-                        weight_limit: WeightLimit::Unlimited,
-                    },
-                    Transact {
-                        origin_type: OriginKind::SovereignAccount,
-                        require_weight_at_most: 8_000_000_000,
-                        call: inner_call.encode().into(),
-                    }
-                ]),
-            ));
-        });
+		Para1::execute_with(|| {
+			let inner_call =
+				sample_runtime::RuntimeCall::System(frame_system::Call::remark_with_event {
+					remark: vec![10],
+				});
 
-        Para2::execute_with(|| {
-            sample_runtime::System::events()
-                .iter()
-                .for_each(|r| println!(">>> {:#?}", r.event));
+			let fee_asset = MultiAsset {
+				id: Concrete(MultiLocation {
+					parents: 1,
+					interior: Junctions::Here,
+				}),
+				fun: Fungible(8_000_000_000),
+			};
 
-            assert!(sample_runtime::System::events().iter().any(|r| matches!(
-                r.event,
-                sample_runtime::RuntimeEvent::System(frame_system::Event::Remarked {
-                    sender: _,
-                    hash: _
-                })
-            )));
-        });
-    }
+			assert_ok!(sample_runtime::PolkadotXcm::send_xcm(
+				Here,
+				MultiLocation::new(1, X1(Parachain(2))),
+				Xcm(vec![
+					WithdrawAsset(fee_asset.clone().into()),
+					BuyExecution {
+						fees: fee_asset.into(),
+						weight_limit: WeightLimit::Unlimited,
+					},
+					Transact {
+						origin_type: OriginKind::SovereignAccount,
+						require_weight_at_most: 8_000_000_000,
+						call: inner_call.encode().into(),
+					}
+				]),
+			));
+		});
 
-    #[test]
-    fn xcm_call_contract() {
-        Network::reset();
+		Para2::execute_with(|| {
+			sample_runtime::System::events()
+				.iter()
+				.for_each(|r| println!(">>> {:#?}", r.event));
 
-        Para1::execute_with(|| {
-            let inner_call =
-                sample_runtime::RuntimeCall::AxelarGateway(axelar_cgp::Call::call_contract {
-                    destination_chain: String::from("ethereum"),
-                    destination_contract_address: String::from(
-                        "0x5f927395213ee6b95de97bddcb1b2b1c0f16844d",
-                    ),
-                    payload: vec![1; 32],
-                });
+			assert!(sample_runtime::System::events().iter().any(|r| matches!(
+				r.event,
+				sample_runtime::RuntimeEvent::System(frame_system::Event::Remarked {
+					sender: _,
+					hash: _
+				})
+			)));
+		});
+	}
 
-            let fee_asset = MultiAsset {
-                id: Concrete(MultiLocation {
-                    parents: 1,
-                    interior: Junctions::Here,
-                }),
-                fun: Fungible(8_000_000_000),
-            };
+	#[test]
+	fn xcm_call_contract() {
+		Network::reset();
 
-            assert_ok!(sample_runtime::PolkadotXcm::send_xcm(
-                Here,
-                MultiLocation::new(1, X1(Junction::Parachain(2))),
-                Xcm(vec![
-                    WithdrawAsset(fee_asset.clone().into()),
-                    BuyExecution {
-                        fees: fee_asset.into(),
-                        weight_limit: WeightLimit::Unlimited,
-                    },
-                    Transact {
-                        origin_type: OriginKind::Xcm,
-                        require_weight_at_most: 8_000_000_000,
-                        call: inner_call.encode().into(),
-                    }
-                ]),
-            ));
-        });
+		Para1::execute_with(|| {
+			let inner_call =
+				sample_runtime::RuntimeCall::AxelarGateway(axelar_cgp::Call::call_contract {
+					destination_chain: String::from("ethereum"),
+					destination_contract_address: String::from(
+						"0x5f927395213ee6b95de97bddcb1b2b1c0f16844d",
+					),
+					payload: vec![1; 32],
+				});
 
-        Para2::execute_with(|| {
-            sample_runtime::System::events()
-                .iter()
-                .for_each(|r| println!(">>> {:#?}", r.event));
+			let fee_asset = MultiAsset {
+				id: Concrete(MultiLocation {
+					parents: 1,
+					interior: Junctions::Here,
+				}),
+				fun: Fungible(8_000_000_000),
+			};
 
-            assert!(sample_runtime::System::events().iter().any(|r| matches!(
-                r.event,
-                sample_runtime::RuntimeEvent::AxelarGateway(axelar_cgp::Event::ContractCall {
-                    sender: _,
-                    destination_chain: _,
-                    destination_contract_address: _,
-                    payload_hash: _,
-                    payload: _
-                })
-            )));
-        });
-    }
+			assert_ok!(sample_runtime::PolkadotXcm::send_xcm(
+				Here,
+				MultiLocation::new(1, X1(Junction::Parachain(2))),
+				Xcm(vec![
+					WithdrawAsset(fee_asset.clone().into()),
+					BuyExecution {
+						fees: fee_asset.into(),
+						weight_limit: WeightLimit::Unlimited,
+					},
+					Transact {
+						origin_type: OriginKind::Xcm,
+						require_weight_at_most: 8_000_000_000,
+						call: inner_call.encode().into(),
+					}
+				]),
+			));
+		});
 
-    #[test]
-    fn call_contract_wrong_origin() {
-        Network::reset();
-        let alice: AccountId = AccountId::from_str(
-            "0xd43593c715fdd31c61141abd04a99fd6822c8558854ccde39a5684e7a56da27d",
-        )
-        .unwrap();
+		Para2::execute_with(|| {
+			sample_runtime::System::events()
+				.iter()
+				.for_each(|r| println!(">>> {:#?}", r.event));
 
-        Para1::execute_with(|| {
-            assert_noop!(
-                sample_runtime::AxelarGateway::call_contract(
-                    RuntimeOrigin::signed(alice),
-                    String::from("ethereum"),
-                    String::from("0x5f927395213ee6b95de97bddcb1b2b1c0f16844d"),
-                    vec![1; 32]
-                ),
-                BadOrigin
-            );
-        });
-    }
+			assert!(sample_runtime::System::events().iter().any(|r| matches!(
+				r.event,
+				sample_runtime::RuntimeEvent::AxelarGateway(axelar_cgp::Event::ContractCall {
+					sender: _,
+					destination_chain: _,
+					destination_contract_address: _,
+					payload_hash: _,
+					payload: _
+				})
+			)));
+		});
+	}
 
-    #[test]
-    fn forward_approved_call_xcm() {
-        Network::reset();
+	#[test]
+	fn call_contract_wrong_origin() {
+		Network::reset();
+		let alice: AccountId = AccountId::from_str(
+			"0xd43593c715fdd31c61141abd04a99fd6822c8558854ccde39a5684e7a56da27d",
+		)
+		.unwrap();
 
-        Para1::execute_with(|| {
-            let alice: AccountId = AccountId::from_str(
-                "0xd43593c715fdd31c61141abd04a99fd6822c8558854ccde39a5684e7a56da27d",
-            )
-            .unwrap();
+		Para1::execute_with(|| {
+			assert_noop!(
+				sample_runtime::AxelarGateway::call_contract(
+					RuntimeOrigin::signed(alice),
+					String::from("ethereum"),
+					String::from("0x5f927395213ee6b95de97bddcb1b2b1c0f16844d"),
+					vec![1; 32]
+				),
+				BadOrigin
+			);
+		});
+	}
 
-            let inner_call = sample_runtime::RuntimeCall::SampleReceiver(
-                sample_receiver::Call::sample_final_call {
-                    data: vec![1, 2, 3],
-                },
-            );
-            let inner_call_bytes = inner_call.encode();
+	#[test]
+	fn forward_approved_call_xcm() {
+		Network::reset();
 
-            let command_id = H256::random();
-            let source_chain = String::from("ethereum");
-            let source_address = String::from("0x5f927395213ee6b95de97bddcb1b2b1c0f16844d");
-            let contract_address = H160::random();
+		Para1::execute_with(|| {
+			let alice: AccountId = AccountId::from_str(
+				"0xd43593c715fdd31c61141abd04a99fd6822c8558854ccde39a5684e7a56da27d",
+			)
+			.unwrap();
 
-            let call_hash = H256::from_slice(&ecdsa::to_eth_signed_message_hash(keccak_256(
-                inner_call_bytes.as_slice(),
-            )));
-            let mut approved_call = command_id.encode();
-            approved_call.append(&mut source_chain.encode());
-            approved_call.append(&mut source_address.encode());
-            approved_call.append(&mut contract_address.encode());
-            approved_call.append(&mut call_hash.encode());
+			let inner_call = sample_runtime::RuntimeCall::SampleReceiver(
+				sample_receiver::Call::sample_final_call {
+					data: vec![1, 2, 3],
+				},
+			);
+			let inner_call_bytes = inner_call.encode();
 
-            let approved_call_hash = H256::from(keccak_256(approved_call.as_slice()));
+			let command_id = H256::random();
+			let source_chain = String::from("ethereum");
+			let source_address = String::from("0x5f927395213ee6b95de97bddcb1b2b1c0f16844d");
+			let contract_address = H160::random();
 
-            axelar_cgp::ContractCallApproved::<sample_runtime::Runtime>::set(
-                approved_call_hash,
-                (),
-            );
-            axelar_cgp::CommandExecuted::<sample_runtime::Runtime>::set(command_id, 2u32);
+			let call_hash = H256::from_slice(&ecdsa::to_eth_signed_message_hash(keccak_256(
+				inner_call_bytes.as_slice(),
+			)));
+			let mut approved_call = command_id.encode();
+			approved_call.append(&mut source_chain.encode());
+			approved_call.append(&mut source_address.encode());
+			approved_call.append(&mut contract_address.encode());
+			approved_call.append(&mut call_hash.encode());
 
-            assert_ok!(sample_runtime::AxelarGateway::forward_approved_call(
-                RuntimeOrigin::signed(alice),
-                command_id,
-                source_chain,
-                source_address,
-                contract_address,
-                inner_call_bytes
-            ));
-        });
+			let approved_call_hash = H256::from(keccak_256(approved_call.as_slice()));
 
-        Para2::execute_with(|| {
-            sample_runtime::System::events()
-                .iter()
-                .for_each(|r| println!(">>> {:#?}", r.event));
+			axelar_cgp::ContractCallApproved::<sample_runtime::Runtime>::set(
+				approved_call_hash,
+				(),
+			);
+			axelar_cgp::CommandExecuted::<sample_runtime::Runtime>::set(command_id, 2u32);
 
-            assert!(sample_runtime::System::events().iter().any(|r| matches!(
-                r.event,
-                sample_runtime::RuntimeEvent::SampleReceiver(
-                    sample_receiver::Event::SampleFinalCall {
-                        sender: _,
-                        data: _,
-                        proxy_chain: _,
-                    }
-                )
-            )));
-        });
-    }
+			assert_ok!(sample_runtime::AxelarGateway::forward_approved_call(
+				RuntimeOrigin::signed(alice),
+				command_id,
+				source_chain,
+				source_address,
+				contract_address,
+				inner_call_bytes
+			));
+		});
 
-    #[test]
-    fn full_execute_flow() {
-        Network::reset();
+		Para2::execute_with(|| {
+			sample_runtime::System::events()
+				.iter()
+				.for_each(|r| println!(">>> {:#?}", r.event));
 
-        let alice: AccountId = AccountId::from_str(
-            "0xd43593c715fdd31c61141abd04a99fd6822c8558854ccde39a5684e7a56da27d",
-        )
-        .unwrap();
+			assert!(sample_runtime::System::events().iter().any(|r| matches!(
+				r.event,
+				sample_runtime::RuntimeEvent::SampleReceiver(
+					sample_receiver::Event::SampleFinalCall {
+						sender: _,
+						data: _,
+						proxy_chain: _,
+					}
+				)
+			)));
+		});
+	}
 
-        Para1::execute_with(|| {
-            // This is the call to be executed in the final parachain
-            let inner_call = sample_runtime::RuntimeCall::SampleReceiver(
-                sample_receiver::Call::sample_final_call {
-                    data: vec![1, 2, 3],
-                },
-            );
-            let inner_call_bytes: Vec<u8> = inner_call.encode();
+	#[test]
+	fn full_execute_flow() {
+		Network::reset();
 
-            /////////// Preparing Execute Batch Payload ///////////
-            let source_chain = String::from("ethereum");
-            let source_address = String::from("0x5f927395213ee6b95de97bddcb1b2b1c0f16844d");
-            let contract_address = H160::random();
-            let payload_hash = H256::from_slice(&ecdsa::to_eth_signed_message_hash(keccak_256(
-                inner_call_bytes.as_slice(),
-            )));
-            let source_tx_hash = H256::random();
-            let source_event_index = U256::from(100);
-            let command_id = H256::random();
+		let alice: AccountId = AccountId::from_str(
+			"0xd43593c715fdd31c61141abd04a99fd6822c8558854ccde39a5684e7a56da27d",
+		)
+		.unwrap();
 
-            // Preparing call to be executed in the Axelar gateway parachain
-            let outer_call = sample_runtime::RuntimeCall::AxelarGateway(
-                axelar_cgp::Call::approve_contract_call {
-                    source_chain: source_chain.clone(),
-                    source_address: source_address.clone(),
-                    contract_address,
-                    payload_hash,
-                    source_tx_hash,
-                    source_event_index,
-                    command_id,
-                },
-            );
+		Para1::execute_with(|| {
+			// This is the call to be executed in the final parachain
+			let inner_call = sample_runtime::RuntimeCall::SampleReceiver(
+				sample_receiver::Call::sample_final_call {
+					data: vec![1, 2, 3],
+				},
+			);
+			let inner_call_bytes: Vec<u8> = inner_call.encode();
 
-            // Preparing sign message, to be signed by the virtual operators
-            let chain_id = 2_u32;
-            let command_x: String = String::from("approveContractCall");
-            let batch_msg: ethabi::Bytes = sample_runtime::AxelarGateway::abi_encode_batch_params(
-                chain_id,
-                vec![command_id.clone()],
-                vec![command_x.clone()],
-                vec![outer_call.clone()],
-            );
-            let sign_msg = ecdsa::to_eth_signed_message_hash(keccak_256(batch_msg.as_slice()));
+			/////////// Preparing Execute Batch Payload ///////////
+			let source_chain = String::from("ethereum");
+			let source_address = String::from("0x5f927395213ee6b95de97bddcb1b2b1c0f16844d");
+			let contract_address = H160::random();
+			let payload_hash = H256::from_slice(&ecdsa::to_eth_signed_message_hash(keccak_256(
+				inner_call_bytes.as_slice(),
+			)));
+			let source_tx_hash = H256::random();
+			let source_event_index = U256::from(100);
+			let command_id = H256::random();
 
-            let operator_0 = ecdsa::generate_keypair();
-            let operator_1 = ecdsa::generate_keypair();
-            let operator_0_public =
-                H160::from(H256::from_slice(keccak_256(&operator_0.0).as_slice()));
-            let operator_1_public =
-                H160::from(H256::from_slice(keccak_256(&operator_1.0).as_slice()));
-            let sig_0 = ecdsa::sign_message(H256::from_slice(&sign_msg), &operator_0.1);
-            let sig_1 = ecdsa::sign_message(H256::from_slice(&sign_msg), &operator_1.1);
-            let proof_bytes = axelar_cgp::utils::proofs::encode(
-                vec![
-                    operator_0_public.to_fixed_bytes(),
-                    operator_1_public.to_fixed_bytes(),
-                ],
-                vec![50u128, 50u128],
-                50u128,
-                vec![sig_0, sig_1],
-            )
-            .to_vec();
+			// Preparing call to be executed in the Axelar gateway parachain
+			let outer_call = sample_runtime::RuntimeCall::AxelarGateway(
+				axelar_cgp::Call::approve_contract_call {
+					source_chain: source_chain.clone(),
+					source_address: source_address.clone(),
+					contract_address,
+					payload_hash,
+					source_tx_hash,
+					source_event_index,
+					command_id,
+				},
+			);
 
-            let op0_addr = Token::Address(operator_0_public.to_fixed_bytes().into())
-                .into_address()
-                .unwrap();
-            let op1_addr = Token::Address(operator_1_public.to_fixed_bytes().into())
-                .into_address()
-                .unwrap();
-            let operators_hash =
-                proof::operators_hash(vec![op0_addr, op1_addr], vec![50u128, 50u128], 50u128);
+			// Preparing sign message, to be signed by the virtual operators
+			let chain_id = 2_u32;
+			let command_x: String = String::from("approveContractCall");
+			let batch_msg: ethabi::Bytes = sample_runtime::AxelarGateway::abi_encode_batch_params(
+				chain_id,
+				vec![command_id.clone()],
+				vec![command_x.clone()],
+				vec![outer_call.clone()],
+			);
+			let sign_msg = ecdsa::to_eth_signed_message_hash(keccak_256(batch_msg.as_slice()));
 
-            // Initialize the Epoch for the above set of operators
-            axelar_cgp::EpochForHash::<sample_runtime::Runtime>::insert(operators_hash, 100);
-            axelar_cgp::CurrentEpoch::<sample_runtime::Runtime>::set(100);
+			let operator_0 = ecdsa::generate_keypair();
+			let operator_1 = ecdsa::generate_keypair();
+			let operator_0_public =
+				H160::from(H256::from_slice(keccak_256(&operator_0.0).as_slice()));
+			let operator_1_public =
+				H160::from(H256::from_slice(keccak_256(&operator_1.0).as_slice()));
+			let sig_0 = ecdsa::sign_message(H256::from_slice(&sign_msg), &operator_0.1);
+			let sig_1 = ecdsa::sign_message(H256::from_slice(&sign_msg), &operator_1.1);
+			let proof_bytes = axelar_cgp::utils::proofs::encode(
+				vec![
+					operator_0_public.to_fixed_bytes(),
+					operator_1_public.to_fixed_bytes(),
+				],
+				vec![50u128, 50u128],
+				50u128,
+				vec![sig_0, sig_1],
+			)
+			.to_vec();
 
-            //////////////////// End Preparing batch Execute /////////////////////
+			let op0_addr = Token::Address(operator_0_public.to_fixed_bytes().into())
+				.into_address()
+				.unwrap();
+			let op1_addr = Token::Address(operator_1_public.to_fixed_bytes().into())
+				.into_address()
+				.unwrap();
+			let operators_hash =
+				proof::operators_hash(vec![op0_addr, op1_addr], vec![50u128, 50u128], 50u128);
 
-            // Axelar Relayer triggers the batch execution
-            assert_ok!(sample_runtime::AxelarGateway::execute(
-                RuntimeOrigin::signed(alice.clone()),
-                proof_bytes.clone(),
-                chain_id,
-                vec![command_id],
-                vec![command_x.clone()],
-                vec![outer_call.clone()]
-            ));
+			// Initialize the Epoch for the above set of operators
+			axelar_cgp::EpochForHash::<sample_runtime::Runtime>::insert(operators_hash, 100);
+			axelar_cgp::CurrentEpoch::<sample_runtime::Runtime>::set(100);
 
-            assert_eq!(
-                axelar_cgp::CommandExecuted::<sample_runtime::Runtime>::get(command_id),
-                chain_id
-            );
-            assert!(sample_runtime::System::events().iter().any(|r| matches!(
-                r.event,
-                sample_runtime::RuntimeEvent::AxelarGateway(
-                    axelar_cgp::Event::ContractCallApproved { .. }
-                )
-            )));
+			//////////////////// End Preparing batch Execute /////////////////////
 
-            // Now that call has been approved, Axelar Relayer forwards the final execution
-            assert_ok!(sample_runtime::AxelarGateway::forward_approved_call(
-                RuntimeOrigin::signed(alice),
-                command_id,
-                source_chain.clone(),
-                source_address.clone(),
-                contract_address,
-                inner_call_bytes
-            ));
+			// Axelar Relayer triggers the batch execution
+			assert_ok!(sample_runtime::AxelarGateway::execute(
+				RuntimeOrigin::signed(alice.clone()),
+				proof_bytes.clone(),
+				chain_id,
+				vec![command_id],
+				vec![command_x.clone()],
+				vec![outer_call.clone()]
+			));
 
-            let mut call_hash_payload = command_id.encode();
-            call_hash_payload.append(&mut source_chain.encode());
-            call_hash_payload.append(&mut source_address.encode());
-            call_hash_payload.append(&mut contract_address.encode());
-            call_hash_payload.append(&mut payload_hash.encode());
+			assert_eq!(
+				axelar_cgp::CommandExecuted::<sample_runtime::Runtime>::get(command_id),
+				chain_id
+			);
+			assert!(sample_runtime::System::events().iter().any(|r| matches!(
+				r.event,
+				sample_runtime::RuntimeEvent::AxelarGateway(
+					axelar_cgp::Event::ContractCallApproved { .. }
+				)
+			)));
 
-            let approved_call_hash = H256::from(keccak_256(call_hash_payload.as_slice()));
+			// Now that call has been approved, Axelar Relayer forwards the final execution
+			assert_ok!(sample_runtime::AxelarGateway::forward_approved_call(
+				RuntimeOrigin::signed(alice),
+				command_id,
+				source_chain.clone(),
+				source_address.clone(),
+				contract_address,
+				inner_call_bytes
+			));
 
-            // After Contract Call approved, the call hash has been removed, so there is no collisions with retrials from the Axelar network
-            assert!(
-                !axelar_cgp::ContractCallApproved::<sample_runtime::Runtime>::contains_key(
-                    approved_call_hash
-                )
-            );
-        });
+			let mut call_hash_payload = command_id.encode();
+			call_hash_payload.append(&mut source_chain.encode());
+			call_hash_payload.append(&mut source_address.encode());
+			call_hash_payload.append(&mut contract_address.encode());
+			call_hash_payload.append(&mut payload_hash.encode());
 
-        Para2::execute_with(|| {
-            sample_runtime::System::events()
-                .iter()
-                .for_each(|r| println!(">>> {:#?}", r.event));
+			let approved_call_hash = H256::from(keccak_256(call_hash_payload.as_slice()));
 
-            // Final call successfully executed across the "wire"
-            assert!(sample_runtime::System::events().iter().any(|r| matches!(
-                r.event,
-                sample_runtime::RuntimeEvent::SampleReceiver(
-                    sample_receiver::Event::SampleFinalCall { .. }
-                )
-            )));
-        });
-    }
+			// After Contract Call approved, the call hash has been removed, so there is no collisions with retrials from the Axelar network
+			assert!(
+				!axelar_cgp::ContractCallApproved::<sample_runtime::Runtime>::contains_key(
+					approved_call_hash
+				)
+			);
+		});
+
+		Para2::execute_with(|| {
+			sample_runtime::System::events()
+				.iter()
+				.for_each(|r| println!(">>> {:#?}", r.event));
+
+			// Final call successfully executed across the "wire"
+			assert!(sample_runtime::System::events().iter().any(|r| matches!(
+				r.event,
+				sample_runtime::RuntimeEvent::SampleReceiver(
+					sample_receiver::Event::SampleFinalCall { .. }
+				)
+			)));
+		});
+	}
 }
