+++ libs/ecdsa/Cargo.toml
@@ -11,9 +11,18 @@ edition = "2021"
 
 [dependencies]
 eth-encode-packed = "0.1.0"
-libsecp256k1 = "0.7"
-rand = "0.8.5"
+libsecp256k1 = {version = "0.7", default-features = false }
+rand = { version = "0.8.5", default-features = false }
 
 # Substrate
-sp-core = { git = "https://github.com/paritytech/substrate", branch = "polkadot-v0.9.33" }
-sp-io = { git = "https://github.com/paritytech/substrate", branch = "polkadot-v0.9.33" }
+sp-core = { git = "https://github.com/paritytech/substrate", default-features = false, branch = "polkadot-v0.9.33" }
+sp-io = { git = "https://github.com/paritytech/substrate", default-features = false, branch = "polkadot-v0.9.33" }
+
+[features]
+default = ['std']
+std = [
+    "sp-core/std",
+    "sp-io/std",
+    "rand/std",
+    "libsecp256k1/std",
+]
